+++ demo/source/demos/advancedRouting/routes.js
@@ -8,15 +8,12 @@ const routes = [
         state: { page: "product" },
         // Set the default tab to details
         match: state => ({ tab: "details", ...state }),
+        // For serialization, remove the tab if it's "details", so /product
+        // is produced whether the tab is specified or not, otherwise
+        // keep the tab the same so the correct child path matches
+        stringify: ({ tab, ...state }) =>
+            tab === "details" ? state : { tab, ...state },
         routes: [
-            {
-                // This route exists so we stringify to the
-                // same url with or without tab details
-                // TODO: Seems a bit inelegant, think of a
-                // way to improve this
-                path: "/",
-                state: { tab: "details" }
-            },
             {
                 path: "/ratings",
                 state: { tab: "ratings" }