+++ demo/source/demos/basicRouting/Root.js
@@ -1,13 +1,13 @@
 import React from "react";
 
-import { SimpleProvider } from "jarl-react";
+import { StateProvider } from "jarl-react";
 
 import Pages from "./Pages";
 
 const Root = ({ history, routes, basePath }) => (
-    <SimpleProvider history={history} routes={routes} basePath={basePath}>
+    <StateProvider history={history} routes={routes} basePath={basePath}>
         <Pages />
-    </SimpleProvider>
+    </StateProvider>
 );
 
 export default Root;