+++ demo/source/demos/codeSplitting/Pages.js
@@ -1,6 +1,5 @@
 import React, { Component } from "react";
 import Helmet from "react-helmet";
-import { compose } from "recompose";
 
 import { routing, Link } from "jarl-react";