+++ packages/jarl-react/source/index.js
@@ -4,4 +4,4 @@ export { default as redirect, Redirect } from "./redirect";
 export { default as Link } from "./Link";
 export { default as Router } from "./Router";
 export { default as routing } from "./routing";
-export { default as SimpleProvider } from "./SimpleProvider";
+export { default as StateProvider } from "./StateProvider";